Move the wabs absorption function into spectra
